音频采样改用float而不是int

+++ src/main/java/lpc/javaTools/media/audio/PCMData.java
@@ -1,10 +1,10 @@
 package lpc.javaTools.media.audio;
 
 public class PCMData {
-	public int[][] samples;
+	public float[][] samples;
 	public int sampleRate;
 	
-	public PCMData(int[][] samples, int sampleRate) {
+	public PCMData(float[][] samples, int sampleRate) {
 		this.samples = samples;
 		this.sampleRate = sampleRate;
 	}
